MBS-8098 Proposal extract DeployStep.kt DeployStep.kt accessible only from `release {}`
